# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Replit Clone API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  clerkId: string;
  username: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  plan?: string;
  projectCount?: number;
  createdAt: string;
}

export interface UserUpdate {
  username?: string;
  displayName?: string;
  bio?: string;
}

export interface Project {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** website, mobile, design, game, slides, document, data-viz, animation */
  type: string;
  /** @nullable */
  language?: string | null;
  userId: string;
  /** @nullable */
  username?: string | null;
  isPublic: boolean;
  isPinned?: boolean;
  /** @nullable */
  subdomain?: string | null;
  forkCount?: number;
  likeCount?: number;
  /** @nullable */
  lastRunAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  type: string;
  language?: string;
  isPublic?: boolean;
  templateId?: string;
}

export interface ProjectUpdate {
  name?: string;
  description?: string;
  isPublic?: boolean;
  isPinned?: boolean;
  subdomain?: string;
}

export interface Template {
  id: string;
  name: string;
  type: string;
  description: string;
  /** @nullable */
  language?: string | null;
  category?: string;
  /** @nullable */
  iconUrl?: string | null;
  popularity?: number;
}

export interface DashboardStats {
  totalProjects: number;
  publicProjects: number;
  totalForks: number;
  recentActivity: number;
}

export interface ActivityItem {
  id: string;
  /** created, updated, forked, published */
  type: string;
  projectId: string;
  projectName: string;
  projectType?: string;
  createdAt: string;
}

export type AgentMessageHistoryItem = {
  role?: string;
  content?: string;
};

export interface AgentMessage {
  message: string;
  projectId: string;
  history?: AgentMessageHistoryItem[];
}

export interface AgentResponse {
  message: string;
  projectId: string;
  suggestions?: string[];
}

export type ListProjectsParams = {
type?: string;
search?: string;
};

export type ListTemplatesParams = {
category?: string;
};

export type ListPublicProjectsParams = {
search?: string;
type?: string;
};