# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Replit Clone API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Project (Repl) operations
  - name: users
    description: User profile operations
  - name: templates
    description: Template operations
  - name: stats
    description: Stats and dashboard operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List user's projects
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /projects/{id}/fork:
    post:
      operationId: forkProject
      tags: [projects]
      summary: Fork a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Forked project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List available templates
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"

  /stats/dashboard:
    get:
      operationId: getDashboardStats
      tags: [stats]
      summary: Get dashboard stats summary
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /stats/activity:
    get:
      operationId: getRecentActivity
      tags: [stats]
      summary: Get recent project activity
      responses:
        "200":
          description: Recent activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /projects/public:
    get:
      operationId: listPublicProjects
      tags: [projects]
      summary: List public/community projects
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Public projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

  /agent/chat:
    post:
      operationId: agentChat
      tags: [projects]
      summary: Send a message to the AI agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentMessage"
      responses:
        "200":
          description: Agent response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    User:
      type: object
      required: [id, clerkId, username, email, createdAt]
      properties:
        id:
          type: string
        clerkId:
          type: string
        username:
          type: string
        email:
          type: string
        displayName:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        plan:
          type: string
        projectCount:
          type: integer
        createdAt:
          type: string

    UserUpdate:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string

    Project:
      type: object
      required: [id, name, type, userId, isPublic, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          description: "website, mobile, design, game, slides, document, data-viz, animation"
        language:
          type: ["string", "null"]
        userId:
          type: string
        username:
          type: ["string", "null"]
        isPublic:
          type: boolean
        isPinned:
          type: boolean
        subdomain:
          type: ["string", "null"]
        forkCount:
          type: integer
        likeCount:
          type: integer
        lastRunAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    ProjectInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        type:
          type: string
        language:
          type: string
        isPublic:
          type: boolean
        templateId:
          type: string

    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isPublic:
          type: boolean
        isPinned:
          type: boolean
        subdomain:
          type: string

    Template:
      type: object
      required: [id, name, type, description]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        description:
          type: string
        language:
          type: ["string", "null"]
        category:
          type: string
        iconUrl:
          type: ["string", "null"]
        popularity:
          type: integer

    DashboardStats:
      type: object
      required: [totalProjects, publicProjects, totalForks, recentActivity]
      properties:
        totalProjects:
          type: integer
        publicProjects:
          type: integer
        totalForks:
          type: integer
        recentActivity:
          type: integer

    ActivityItem:
      type: object
      required: [id, type, projectId, projectName, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          description: "created, updated, forked, published"
        projectId:
          type: string
        projectName:
          type: string
        projectType:
          type: string
        createdAt:
          type: string

    AgentMessage:
      type: object
      required: [message, projectId]
      properties:
        message:
          type: string
        projectId:
          type: string
        history:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              content:
                type: string

    AgentResponse:
      type: object
      required: [message, projectId]
      properties:
        message:
          type: string
        projectId:
          type: string
        suggestions:
          type: array
          items:
            type: string