# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Replit Clone API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: string;
  /** created, updated, forked, published */
  type: string;
  projectId: string;
  projectName: string;
  projectType?: string;
  createdAt: string;
}