# lib/api-zod/src/generated/types/agentMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Replit Clone API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentMessageHistoryItem } from './agentMessageHistoryItem';

export interface AgentMessage {
  message: string;
  projectId: string;
  history?: AgentMessageHistoryItem[];
}