# lib/api-zod/src/generated/types/agentResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Replit Clone API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentResponse {
  message: string;
  projectId: string;
  suggestions?: string[];
}