# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Replit Clone API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalProjects: number;
  publicProjects: number;
  totalForks: number;
  recentActivity: number;
}