# lib/api-zod/src/generated/types/projectInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Replit Clone API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  type: string;
  language?: string;
  isPublic?: boolean;
  templateId?: string;
}