# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Replit Clone API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectUpdate {
  name?: string;
  description?: string;
  isPublic?: boolean;
  isPinned?: boolean;
  subdomain?: string;
}