# lib/api-zod/src/generated/types/template.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Replit Clone API
 * OpenAPI spec version: 0.1.0
 */

export interface Template {
  id: string;
  name: string;
  type: string;
  description: string;
  /** @nullable */
  language?: string | null;
  category?: string;
  /** @nullable */
  iconUrl?: string | null;
  popularity?: number;
}